backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.6.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
fasttext==0.9.1
flake8==3.7.9
idna==2.9
joblib==0.14.1
JPype1==0.7.1
kiwisolver==1.1.0
konlpy==0.5.2
lxml==4.5.0
matplotlib==3.2.0rc3
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
pathspec==0.7.0
Pillow==7.0.0
pybind11==2.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.1.8
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
toml==0.10.0
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.8
